refactor(backend): decouple jwt from auth domain service
Moved JWT verification logic out of `AuthService` into a new `SupabaseJWTVerifier` that implements a `TokenVerifierProtocol` to strictly adhere to Clean Architecture principles.

- Added `TokenVerifierProtocol` in domain interfaces.
- Implemented `SupabaseJWTVerifier` in `app/infrastructure/auth/jwt_verifier.py`.
- Refactored `AuthService` to accept `TokenVerifierProtocol` via constructor injection.
- Updated FastAPI dependency injections to provide the new `TokenVerifierProtocol` implementation.
- Refactored `websocket.py` to utilize the dependency graph accessor for token validation.
- Updated all related unit tests.
- Replaced `...` with `pass` in abstract Protocol classes to satisfy CodeQL.
- Migrated external blocking library calls (like PyJWKClient token fetching) into `asyncio.to_thread` to prevent blocking the async event loop.
